refactor: added global types
