+++ src/app/model/store.ts
@@ -1,6 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { setupListeners } from "@reduxjs/toolkit/query";
-import { useSelector, TypedUseSelectorHook, useDispatch } from "react-redux";
 import { meetingsApi, meetingReducer } from "@/entities/meeting";
 import { slotReducer, slotsApi } from "@/entities/slot";
 import { usersReducer, usersApi } from "@/entities/specialist";
@@ -34,7 +33,3 @@ export const store = configureStore({
       .prepend(listenerMiddlewareAuth.middleware),
 });
 setupListeners(store.dispatch);
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
-export const useAppDispatch = () => useDispatch<AppDispatch>();
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
