+++ src/app/model/types.d.ts
@@ -1,0 +1,7 @@
+import { store } from "./store";
+declare global {
+  declare type RootState = ReturnType<typeof store.getState>;
+  declare type AppDispatch = typeof store.dispatch;
+}
+
+export {};
