+++ src/app/ui/index.tsx
@@ -2,15 +2,14 @@ import "./index.css";
 import { routes } from "../router";
 import { RouterProvider } from "react-router-dom";
 import { Suspense } from "react";
-import { Provider } from "react-redux";
-import { store } from "@/entities/root";
+import { AppStoreProvider } from "../providers";
 
 export function App() {
   return (
-    <Provider store={store}>
+    <AppStoreProvider>
       <Suspense fallback={<div>Заагрузка...</div>}>
         <RouterProvider router={routes} />
       </Suspense>
-    </Provider>
+    </AppStoreProvider>
   );
 }
