+++ src/entities/material/model/selectors.ts
@@ -1,5 +1,4 @@
 import { createSelector } from "@reduxjs/toolkit";
-import { RootState } from "@/entities/root";
 
 export const getMaterials = (state: RootState) => state.materials;
