+++ src/entities/meeting/model/selectors.ts
@@ -1,5 +1,4 @@
 import { createSelector } from "@reduxjs/toolkit";
-import { RootState } from "@/entities/root";
 
 export const getDay = (state: RootState) => state.day;
