--- src/entities/root/index.ts
@@ -1,6 +1,0 @@
-export {
-  type RootState,
-  useAppDispatch,
-  useAppSelector,
-  store,
-} from "./model/store";
