+++ src/entities/slot/model/selectors.ts
@@ -1,2 +1,1 @@
-import { RootState } from "@/entities/root";
 export const getSlot = (state: RootState) => state.slot;
