+++ src/entities/specialist/model/selectors.ts
@@ -1,5 +1,5 @@
 import { createSelector } from "@reduxjs/toolkit";
-import { RootState } from "@/entities/root";
+
 export const getRecord = (state: RootState) => state.users;
 export const getChosenOptions = createSelector(getRecord, (record) => {
   return record.chosenOptions;
