+++ src/shared/lib/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useModal } from "./useModal";
 export { useDebounce } from "./useDebounce";
 export { useSuggestions } from "./useSuggestions";
+export { useAppDispatch, useAppSelector } from "./useRedux";
