+++ src/shared/lib/hooks/useRedux.ts
@@ -1,0 +1,4 @@
+import { TypedUseSelectorHook, useSelector, useDispatch } from "react-redux";
+
+export const useAppDispatch = () => useDispatch<AppDispatch>();
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
